+++ components/calendar/src/stories/calendar-input.prod.stories.js
@@ -26,19 +26,18 @@ export default {
 
 const buildCalendar =
     ({ date, locale, calendar }) =>
-    () =>
-        (
-            <CalendarStoryWrapper
-                component={CalendarInput}
-                dir="ltr"
-                timeZone="Africa/Khartoum"
-                weekDayFormat="short"
-                date={date}
-                locale={locale}
-                calendar={calendar}
-                onDateSelect={() => {}}
-            />
-        )
+    () => (
+        <CalendarStoryWrapper
+            component={CalendarInput}
+            dir="ltr"
+            timeZone="Africa/Khartoum"
+            weekDayFormat="short"
+            date={date}
+            locale={locale}
+            calendar={calendar}
+            onDateSelect={() => {}}
+        />
+    )
 
 export const EthiopicWithAmharic = buildCalendar({
     calendar: 'ethiopic',